requirements.txt
colour
pyyaml
requests